afmongodb: remove unused last_msg_stamp
The mongodb driver kept track of the timestamp of the last message
delivered. This seems to have been the remains of code taken from
the affile driver and has never been used.

It is originally used to assist in the cleaning up of destination
writers.

By getting rid of this, we can get rid of the supporting functionality
from logthrdstdrv.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>